@@ -2,73 +2,57 @@ using AutoMapper;
 using Intecgra.Cerberus.Api.Extensions;
 using Intecgra.Cerberus.Api.Filters;
 using Intecgra.Cerberus.Domain.Config;
-using Intecgra.Cerberus.Infrastructure.Data;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc.Authorization;
-using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.OpenApi.Models;
 
-namespace Intecgra.Cerberus.Api
+namespace Intecgra.Cerberus.Api;
+
+public class Startup
 {
-    public class Startup
+    public Startup(IConfiguration configuration)
     {
-        public Startup(IConfiguration configuration)
+        Configuration = configuration;
+    }
+
+    private IConfiguration Configuration { get; }
+
+    // This method gets called by the runtime. Use this method to add services to the container.
+    public void ConfigureServices(IServiceCollection services)
+    {
+        var mapConfig = new MapperConfiguration(mc => mc.AddProfile(new MappingProfile()));
+        var mapper = mapConfig.CreateMapper();
+        services.AddSingleton(mapper);
+
+        services.AddControllers(cfg => { cfg.Filters.Add(typeof(AppExceptionFilter)); });
+
+        services.AddAppStore();
+
+        services.AddSwaggerGen(c =>
         {
-            Configuration = configuration;
+            c.SwaggerDoc("v1", new OpenApiInfo {Title = "Intecgra.Cerberus.Api", Version = "v1"});
+        });
+        services.AddSingleton(_ => Configuration);
+    }
+
+    // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
+    public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+    {
+        if (env.IsDevelopment())
+        {
+            app.UseDeveloperExceptionPage();
+            app.UseSwagger();
+            app.UseSwaggerUI(c => c.SwaggerEndpoint("/swagger/v1/swagger.json", "Intecgra.Cerberus.Api v1"));
         }
 
-        private IConfiguration Configuration { get; }
+        app.UseAuthentication();
+        app.UseAuthorization();
 
-        // This method gets called by the runtime. Use this method to add services to the container.
-        public void ConfigureServices(IServiceCollection services)
-        {
-            var mapConfig = new MapperConfiguration(mc => mc.AddProfile(new MappingProfile()));
-            var mapper = mapConfig.CreateMapper();
-            services.AddSingleton(mapper);
-
-            services.AddControllers(cfg => { cfg.Filters.Add(typeof(AppExceptionFilter)); });
-
-            services.AddAuthentication(options =>
-            {
-                options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-                options.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-            }).AddJwtBearer(options =>
-            {
-                options.Authority = Configuration["Authentication:Domain"];
-                options.Audience = Configuration["Authentication:Audience"];
-            });
-
-            services.AddAppStore();
-
-            services.AddSwaggerGen(c =>
-            {
-                c.SwaggerDoc("v1", new OpenApiInfo {Title = "Intecgra.Cerberus.Api", Version = "v1"});
-            });
-            services.AddSingleton(_ => Configuration);
-        }
-
-        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
-        {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-                app.UseSwagger();
-                app.UseSwaggerUI(c => c.SwaggerEndpoint("/swagger/v1/swagger.json", "Intecgra.Cerberus.Api v1"));
-            }
-
-            app.UseAuthentication();
-            app.UseAuthorization();
-
-            app.UseRouting();
-            app.UseAuthorization();
-            app.UseEndpoints(endpoints => { endpoints.MapControllers(); });
-        }
+        app.UseRouting();
+        app.UseAuthorization();
+        app.UseEndpoints(endpoints => { endpoints.MapControllers(); });
     }
 }
